refactor(submission): decouple SubmissionService and SubmissionSubscriptionService
